backend/MyProject/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
pillow==11.1.0
promise==2.3
psycopg==3.2.6
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
razorpay==1.4.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.0
urllib3==2.3.0
dnspython
uvicorn
django-environ==0.9.0
django-ratelimit==3.0.1
argon2-cffi==23.1.0
channels-redis==4.1.0
redis==4.6.0